Update to run whole app with docker
The local frontend was still hitting the real backend due to a hardcoded
URL for the api, this removes that. Needed to set up a proxy for chrome
to not show a CORS error between node dev server and the api server. The
only remaining weird thing to work out is it still hits aws for the
images, there's probably a way to have node serve them or use a fake aws
server, but not actively developing so not really worth it.